pyproject.toml
[project]
name = "wetterstation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "mariadb>=1.1.13",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.9",
]
